Fill `ribasim_testmodels` and let `python` and `api` depend on it
