+++ python/ribasim_testmodels/ribasim_testmodels/__init__.py
@@ -1,5 +1,9 @@
 __version__ = "0.1.1"
 
-import ribasim
+from ribasim_testmodels.basic import (
+    basic_model,
+    basic_transient_model,
+    tabulated_rating_curve_model,
+)
 
-__all__ = ["ribasim"]
+__all__ = ["basic_model", "basic_transient_model", "tabulated_rating_curve_model"]
